Refactor placeShip method.
Refactored placeShip method to avoid putting ship out of the game board
and overlapping other ships.
Added checkCoords function to the coordinates where the ship will be
placed.
It checks the coordinates if it's still inside the game board and it
also check if there are ships on the said coordinates.
Added coordsOK variable which calls on the checkCoords function; If the
said variable has a value of false, it prevents a ship to be places on
the requested coordinates.
Placed back the placedShip increment logic on the placeShip method for
it to be processed only when a ship is placed.
Removed the logic in the spawnCoords function which divide the ship and
put the other part on the next row when it's length is over the board.