Add attacking mechanism to the gameboard module
Created an allShips array containing all the ship; The said array will
be used for the attacking mechanism.
Created receiveAttack method which receive a coordinate as an argument.
The receiveAttack method is used to facilitate the attack on grid of the
game board.
Created the attackShip function that will be used to call the ship's hit
function when a ship is shot.