Add methods to player factory
Created a variable "turn" which stores the value of the player's turn.
Created showTurn method which returns the "turn" value.
Created attackBoard method for attacking the board of the enemy if it's
the player's turn, and for starting the "turn" of the enemy.
Created the startTurn method to which makes the player's turn variable
to true.