Refactor A.I. module and tests
Added turn variable which holds the value of the A.I.'s turn.
Refactored enterCoords method and added the player's turn as parameter.
Added a logic which turns the A.I.'s turn into false on the said method.
The enterCoords will now call the player's turn after making his turn.
Created the showTurn method to show the A.I. turn value.
Created the startTurn method to change the A.I.'s turn value into true.
The test of not repeating coordinates were reduced to 9 enterCoords
calls due to the error of calling the jest function for more than 9.