Add logic to announce the winner of the game.
Retracted the commenting out of showMessage for announcing the winner of
the game.
Refactored it to announce the winner on the next call after announcing
that all ship has been sunk on a board.
Add a winner variable to store the data of the winner.
Add a boolean variable callWinner for using as a factor to announce the
winner on the showMessage method.